fix frontmatter script
